// frontend/core-ui/types/Organisation.ts
export type CoreOrganisationProps = {
  parent: string | null
  slug: string | null
  primary_maintainer?: string | null
  name: string
  ror_id: string | null
  is_tenant: boolean
  website: string | null
  logo_id: string | null
}

// organisation colums used in editOrganisation.updateOrganisation
// NOTE! update when type Organisation changes
export const columsForUpdate = [
  'id',
  'description',
  ...columsForCreate
]
export type Organisation = CoreOrganisationProps & {
  id: string | null
  // about page content created by maintainer
  description: string | null
  parent_names?: string
  rsd_path?: string
}
// ...
export type OrganisationForOverview = Organisation & {
  id: string
  slug: string
  logo_id: string | null
  software_cnt: number | null
  project_cnt: number | null
  children_cnt: number | null
  release_cnt: number | null
  rsd_path: string
}